@@ -718,8 +718,6 @@ public class CraftEventFactory {
         org.bukkit.entity.AnimalTamer bukkitTamer = (tamer != null ? tamer.getBukkitEntity() : null);
         CraftServer craftServer = (CraftServer) bukkitEntity.getServer();
 
-        entity.setPersistenceRequired(true);
-
         EntityTameEvent event = new EntityTameEvent((LivingEntity) bukkitEntity, bukkitTamer);
         craftServer.getPluginManager().callEvent(event);
         return event;
@@ -929,6 +927,8 @@ public class CraftEventFactory {
                 }
             } else if ("thorns".equals(source.msgId)) {
                 cause = DamageCause.THORNS;
+            } else if ("sonic_boom".equals(source.msgId)) {
+                cause = DamageCause.SONIC_BOOM;
             }
 
             return callEntityDamageEvent(damager, entity, cause, modifiers, modifierFunctions, cancelled);
