@@ -8,21 +8,18 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
 import net.minecraft.core.BlockPosition;
 import net.minecraft.core.Holder;
-import net.minecraft.core.IRegistry;
 import net.minecraft.core.IRegistryCustom;
 import net.minecraft.server.level.RegionLimitedWorldAccess;
 import net.minecraft.server.level.WorldServer;
 import net.minecraft.util.random.WeightedRandomList;
 import net.minecraft.world.entity.EnumCreatureType;
 import net.minecraft.world.level.BlockColumn;
-import net.minecraft.world.level.ChunkCoordIntPair;
 import net.minecraft.world.level.GeneratorAccessSeed;
 import net.minecraft.world.level.LevelHeightAccessor;
 import net.minecraft.world.level.StructureManager;
 import net.minecraft.world.level.biome.BiomeBase;
 import net.minecraft.world.level.biome.BiomeManager;
 import net.minecraft.world.level.biome.BiomeSettingsMobs;
-import net.minecraft.world.level.biome.Climate;
 import net.minecraft.world.level.biome.WorldChunkManager;
 import net.minecraft.world.level.block.ITileEntity;
 import net.minecraft.world.level.block.entity.TileEntity;
@@ -32,13 +29,15 @@ import net.minecraft.world.level.chunk.IChunkAccess;
 import net.minecraft.world.level.chunk.ProtoChunk;
 import net.minecraft.world.level.levelgen.HeightMap;
 import net.minecraft.world.level.levelgen.LegacyRandomSource;
+import net.minecraft.world.level.levelgen.RandomState;
 import net.minecraft.world.level.levelgen.SeededRandom;
 import net.minecraft.world.level.levelgen.WorldGenStage;
 import net.minecraft.world.level.levelgen.blending.Blender;
-import net.minecraft.world.level.levelgen.structure.templatesystem.DefinedStructureManager;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplateManager;
 import org.bukkit.block.Biome;
 import org.bukkit.craftbukkit.CraftHeightMap;
 import org.bukkit.craftbukkit.block.CraftBlock;
+import org.bukkit.craftbukkit.util.RandomSourceWrapper;
 import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.generator.ChunkGenerator.BiomeGrid;
 import org.bukkit.generator.ChunkGenerator.ChunkData;
@@ -93,13 +92,12 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
         this.generator = generator;
     }
 
-    private static SeededRandom getSeededRandom() {
-        return new SeededRandom(new LegacyRandomSource(0));
+    public net.minecraft.world.level.chunk.ChunkGenerator getDelegate() {
+        return delegate;
     }
 
-    @Override
-    public net.minecraft.world.level.chunk.ChunkGenerator withSeed(long i) {
-        return new CustomChunkGenerator(this.world, delegate.withSeed(i), this.generator);
+    private static SeededRandom getSeededRandom() {
+        return new SeededRandom(new LegacyRandomSource(0));
     }
 
     @Override
@@ -118,16 +116,16 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
     }
 
     @Override
-    public void createStructures(IRegistryCustom iregistrycustom, StructureManager structuremanager, IChunkAccess ichunkaccess, DefinedStructureManager definedstructuremanager, long i) {
+    public void createStructures(IRegistryCustom iregistrycustom, RandomState randomstate, StructureManager structuremanager, IChunkAccess ichunkaccess, StructureTemplateManager structuretemplatemanager, long i) {
         if (generator.shouldGenerateStructures()) {
-            super.createStructures(iregistrycustom, structuremanager, ichunkaccess, definedstructuremanager, i);
+            super.createStructures(iregistrycustom, randomstate, structuremanager, ichunkaccess, structuretemplatemanager, i);
         }
     }
 
     @Override
-    public void buildSurface(RegionLimitedWorldAccess regionlimitedworldaccess, StructureManager structuremanager, IChunkAccess ichunkaccess) {
+    public void buildSurface(RegionLimitedWorldAccess regionlimitedworldaccess, StructureManager structuremanager, RandomState randomstate, IChunkAccess ichunkaccess) {
         if (generator.shouldGenerateSurface()) {
-            delegate.buildSurface(regionlimitedworldaccess, structuremanager, ichunkaccess);
+            delegate.buildSurface(regionlimitedworldaccess, structuremanager, randomstate, ichunkaccess);
         }
 
         CraftChunkData chunkData = new CraftChunkData(this.world.getWorld(), ichunkaccess);
@@ -135,7 +133,7 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
         int x = ichunkaccess.getPos().x;
         int z = ichunkaccess.getPos().z;
         random.setSeed((long) x * 341873128712L + (long) z * 132897987541L);
-        generator.generateSurface(this.world.getWorld(), random, x, z, chunkData);
+        generator.generateSurface(this.world.getWorld(), new RandomSourceWrapper.RandomWrapper(random), x, z, chunkData);
 
         if (generator.shouldGenerateBedrock()) {
             random = getSeededRandom();
@@ -145,7 +143,7 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
 
         random = getSeededRandom();
         random.setSeed((long) x * 341873128712L + (long) z * 132897987541L);
-        generator.generateBedrock(this.world.getWorld(), random, x, z, chunkData);
+        generator.generateBedrock(this.world.getWorld(), new RandomSourceWrapper.RandomWrapper(random), x, z, chunkData);
         chunkData.breakLink();
 
         // return if new api is used
@@ -223,9 +221,9 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
     }
 
     @Override
-    public void applyCarvers(RegionLimitedWorldAccess regionlimitedworldaccess, long seed, BiomeManager biomemanager, StructureManager structuremanager, IChunkAccess ichunkaccess, WorldGenStage.Features worldgenstage_features) {
+    public void applyCarvers(RegionLimitedWorldAccess regionlimitedworldaccess, long seed, RandomState randomstate, BiomeManager biomemanager, StructureManager structuremanager, IChunkAccess ichunkaccess, WorldGenStage.Features worldgenstage_features) {
         if (generator.shouldGenerateCaves()) {
-            delegate.applyCarvers(regionlimitedworldaccess, seed, biomemanager, structuremanager, ichunkaccess, worldgenstage_features);
+            delegate.applyCarvers(regionlimitedworldaccess, seed, randomstate, biomemanager, structuremanager, ichunkaccess, worldgenstage_features);
         }
 
         if (worldgenstage_features == WorldGenStage.Features.LIQUID) { // stage check ensures that the method is only called once
@@ -235,16 +233,16 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
             int z = ichunkaccess.getPos().z;
             random.setDecorationSeed(seed, 0, 0);
 
-            generator.generateCaves(this.world.getWorld(), random, x, z, chunkData);
+            generator.generateCaves(this.world.getWorld(), new RandomSourceWrapper.RandomWrapper(random), x, z, chunkData);
             chunkData.breakLink();
         }
     }
 
     @Override
-    public CompletableFuture<IChunkAccess> fillFromNoise(Executor executor, Blender blender, StructureManager structuremanager, IChunkAccess ichunkaccess) {
+    public CompletableFuture<IChunkAccess> fillFromNoise(Executor executor, Blender blender, RandomState randomstate, StructureManager structuremanager, IChunkAccess ichunkaccess) {
         CompletableFuture<IChunkAccess> future = null;
         if (generator.shouldGenerateNoise()) {
-            future = delegate.fillFromNoise(executor, blender, structuremanager, ichunkaccess);
+            future = delegate.fillFromNoise(executor, blender, randomstate, structuremanager, ichunkaccess);
         }
 
         java.util.function.Function<IChunkAccess, IChunkAccess> function = (ichunkaccess1) -> {
@@ -254,7 +252,7 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
             int z = ichunkaccess1.getPos().z;
             random.setSeed((long) x * 341873128712L + (long) z * 132897987541L);
 
-            generator.generateNoise(this.world.getWorld(), random, x, z, chunkData);
+            generator.generateNoise(this.world.getWorld(), new RandomSourceWrapper.RandomWrapper(random), x, z, chunkData);
             chunkData.breakLink();
             return ichunkaccess1;
         };
@@ -263,7 +261,7 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
     }
 
     @Override
-    public int getBaseHeight(int i, int j, HeightMap.Type heightmap_type, LevelHeightAccessor levelheightaccessor) {
+    public int getBaseHeight(int i, int j, HeightMap.Type heightmap_type, LevelHeightAccessor levelheightaccessor, RandomState randomstate) {
         if (implementBaseHeight) {
             try {
                 SeededRandom random = getSeededRandom();
@@ -271,13 +269,13 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
                 int zChunk = j >> 4;
                 random.setSeed((long) xChunk * 341873128712L + (long) zChunk * 132897987541L);
 
-                return generator.getBaseHeight(this.world.getWorld(), random, i, j, CraftHeightMap.fromNMS(heightmap_type));
+                return generator.getBaseHeight(this.world.getWorld(), new RandomSourceWrapper.RandomWrapper(random), i, j, CraftHeightMap.fromNMS(heightmap_type));
             } catch (UnsupportedOperationException exception) {
                 implementBaseHeight = false;
             }
         }
 
-        return delegate.getBaseHeight(i, j, heightmap_type, levelheightaccessor);
+        return delegate.getBaseHeight(i, j, heightmap_type, levelheightaccessor, randomstate);
     }
 
     @Override
@@ -291,8 +289,8 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
     }
 
     @Override
-    public void addDebugScreenInfo(List<String> list, BlockPosition blockposition) {
-        delegate.addDebugScreenInfo(list, blockposition);
+    public void addDebugScreenInfo(List<String> list, RandomState randomstate, BlockPosition blockposition) {
+        delegate.addDebugScreenInfo(list, randomstate, blockposition);
     }
 
     @Override
@@ -313,13 +311,8 @@ public class CustomChunkGenerator extends InternalChunkGenerator {
     }
 
     @Override
-    public BlockColumn getBaseColumn(int i, int j, LevelHeightAccessor levelheightaccessor) {
-        return delegate.getBaseColumn(i, j, levelheightaccessor);
-    }
-
-    @Override
-    public Climate.Sampler climateSampler() {
-        return delegate.climateSampler();
+    public BlockColumn getBaseColumn(int i, int j, LevelHeightAccessor levelheightaccessor, RandomState randomstate) {
+        return delegate.getBaseColumn(i, j, levelheightaccessor, randomstate);
     }
 
     @Override
